Streamline project for Railway deployment
