+++ .env.example
@@ -1,0 +1,1 @@
+TELEGRAM_TOKEN=your_bot_token_here
