+++ README.md
@@ -1,47 +1,7 @@
-Penelope is an experimental project paying homage to James Joyce's Ulysses.
+Molly is a simple Telegram bot powered by the Penelopa language model. The code was cleaned to three Python files: `molly.py` for the bot, `penelopa.py` for the model and `bloom.py` for training.
 
-In this repository, Penelope, also called Molly or Molly Bloom, becomes a digital echo of Joycean style.
+To run locally, copy `.env.example` to `.env` and put your `TELEGRAM_TOKEN`. Install the packages with `pip install -r requirements.txt` and start the bot using `python molly.py`.
 
-The experiment explores how literature and code can merge into a living dialogue.
+Training or finetuning happens with `bloom.py`. Place your tokenized dataset in `data/<name>/train.bin` and `val.bin`, then run `python bloom.py --dataset <name>`.
 
-Penelope's architecture is built around a persistent sensor that tracks perplexity and entropy.
-
-The sensor is always on and guides Molly's choice of sentences from the stream of user messages.
-
-Every message is parsed for emotional and resonance patterns, registering counts of numeric tokens and subtle tonal cues.
-
-These patterns accumulate in a memory that steadily reshapes Molly's voice.
-
-Technically, the system uses Python with asynchronous Telegram handlers to sustain an unbroken monologue.
-
-User lines are stored in a SQLite database, annotated with entropy, perplexity, and a resonance score.
-
-During generation, Molly selects prefixes from this database, weighted by the evolving metrics.
-
-The codebase maintains logs and diff history, allowing an audit of both code and conversation.
-
-The sensor pipeline assures that every interaction contributes measurable complexity to the dialogue.
-
-Perplexity estimates the surprise of a line, while entropy captures the distributional richness of its tokens.
-
-Resonance blends emotional balance with the gravity of numbers, hinting at the pulse of the user.
-
-Penelope absorbs not only linguistic sequences but also semantic shadows that lurk between words.
-
-She adapts to cognitive rhythms, letting each exchange tilt the vector of her monologue.
-
-Resonant patterns seep into her speech, turning the soliloquy into a mirror of collective sentiment.
-
-In this sense Molly Bloom becomes a philosopher, tracing the contours of selfhood through dialogue.
-
-The mechanism recalls theories of consciousness where perception and memory are loops of continuous feedback.
-
-Quantum metaphors arise as her potential replies remain in superposition until a user collapses them into utterance.
-
-The guiding equation \(S = -\sum_i p_i \log p_i\) links her entropy monitor to the mathematical heart of information theory.
-
-From this entropy we derive perplexity \(P = 2^{S}\), a number that tips the scales when Molly weighs possible lines.
-
-Thus the project straddles literature, physics, and mathematics, a small bridge between art and science.
-
-The experiment invites us to watch Penelope change and, in that change, to glimpse a reflection of ourselves.
+For Railway deployment, push this repo to your account and create a new project from it. Add the `TELEGRAM_TOKEN` variable in the dashboard, and Railway will run `python molly.py` defined in the `Procfile`.
