+++ requirements.txt
@@ -1,0 +1,5 @@
+python-telegram-bot
+python-dotenv
+numpy
+torch
+tiktoken
